% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioNetwork.R
\name{bioNetwork}
\alias{bioNetwork}
\title{BioNetWork Function}
\usage{
bioNetwork(File = NULL, upload2 = NULL, phy = FALSE, layOut = 1,
  package = FALSE)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows you to analyze genes from a signature using the bionet algorithm from the package "bionet" and uses interactome dataset
}
\examples{
bioNetwork() will analyze network for vorinostat signature from ilincs
}
\keyword{bioNetwork}
