% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3bioNetwork.R
\name{D3bioNetwork}
\alias{D3bioNetwork}
\title{D3bioNetWork Function}
\usage{
D3bioNetwork(File = NULL, upload4 = NULL)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows you to analyze genes from a signature using the bionet algorithm from the package "bionet" and uses interactome dataset.Visualization done using D3.js
}
\examples{
D3bioNetwork() will analyze network for vorinostat signature from ilincs with D3.js visualization
}
\keyword{D3bioNetwork}
