% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3layout.R
\name{D3layout}
\alias{D3layout}
\title{D3 Top Hundred Network Function}
\usage{
D3layout(File = NULL, upload3 = NULL)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows you to analyze top 100 genes from a signature using the interactome dataset using D3.js
}
\examples{
D3layout()  will analyze network for vorinostat signature from ilincs visualized using D3.js
}
\keyword{topHundredNetwork}
