% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmGWAS.R
\name{dmGWAS}
\alias{dmGWAS}
\title{dmGWAS Function}
\usage{
dmGWAS(File = NULL, upload3 = NULL, layOut = 1, proteinN = 2,
  phy = FALSE, enrich = NULL, package = FALSE)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows you to analyze genes  using  algorithm mentioned in Ideker et.al. 2002
}
\examples{
dmGWAS()  will analyze network for genes in signature from ilincs using ideker et. al. 2002 algorithm
}
\keyword{dmGWAS}
