% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfoFromPortals.R
\name{geneInfoFromPortals}
\alias{geneInfoFromPortals}
\title{Gene Information from Portal Function}
\usage{
geneInfoFromPortals(geneList, symbol = T, names = F)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows to extract gene names corresponding to gene names
}
\examples{
geneInfoFromPortals() converts gene IDs to gene names
}
\keyword{GeneInfoFromPortals}
