% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedBioNetwork.R
\name{modifiedBioNetwork}
\alias{modifiedBioNetwork}
\title{ModifiedBioNetWork Function}
\usage{
modifiedBioNetwork(File = NULL, upload4 = NULL, phy = FALSE, layOut = 1,
  package = FALSE)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows you to analyze genes from a signature using the bionet algorithm from the package "bionet" and uses interactome dataset
}
\examples{
modifiedBioNetwork() will analyze network for vorinostat signature from ilincs
}
\keyword{modifiedbioNetwork}
