% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR.R
\name{RWR}
\alias{RWR}
\title{RWR Function}
\usage{
RWR(File = NULL, phy = FALSE, layOut = 1, package = FALSE,
  nodeGoData = NULL, edgeGoData = NULL, proteinN = 1)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows you to analyze genes from a signature using the random walk with restart algorithm and uses STRING dataset
}
\examples{
RWR() will analyze network for vorinostat signature from ilincs
}
\keyword{RWR}
