% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortNetwork.R
\name{sortNetwork}
\alias{sortNetwork}
\title{Sort Network Function}
\usage{
sortNetwork(x)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows you to sort top 100 genes from a signature
}
\examples{
sortNetwork()  will sort network for any signature file
}
\keyword{sortNetwork}
