% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topHundredNetwork.R
\name{topHundredNetwork}
\alias{topHundredNetwork}
\title{Top Hundred Network Function}
\usage{
topHundredNetwork(File = NULL, phy = FALSE, layOut = 1, package = FALSE,
  nodeGoData = NULL, edgeGoData = NULL, proteinN = 1)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows you to analyze top 100 genes from a signature using the interactome dataset
}
\examples{
topHundredNetwork()  will analyze network for vorinostat signature from ilincs
}
\keyword{topHundredNetwork}
