% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topHundredNetwork.R
\name{topHundredNetwork}
\alias{topHundredNetwork}
\title{Top Hundred Network Function}
\usage{
topHundredNetwork(File = NULL, upload1 = NULL, layOut = 1, proteinN = 2,
  phy = FALSE, enrich = NULL, package = FALSE)
}
\arguments{
\item{File}{? Defaults to NULL.}
}
\description{
This function allows you to analyze top 100 genes from a signature using the interactome dataset
}
\examples{
topHundredNetwork()  will analyze network for vorinostat signature from ilincs
}
\keyword{topHundredNetwork}
